+++ src/main/java/com/google/netpcapanalysis/interfaces/dao/PCAPDao.java
@@ -8,10 +8,7 @@ public interface PCAPDao {
 
     public ArrayList<PCAPdata> getPCAPObjects(String searchEntity);
     public void setPCAPObjects(ArrayList<PCAPdata>  data, String searchEntity);
-
-    public ArrayList<PCAPdata> getUniqueIPs(ArrayList<PCAPdata> allData);
-    public String findMyIP(ArrayList<PCAPdata> allData);
-
+    
     public String searchMaliciousDB(String seachIP);
     public void setMaliciousIPObjects(MaliciousPacket data);
 }
